requirements.txt
fastapi
uvicorn
pydantic-settings
motor
spacy